backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# OpenAI & LangChain
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.19
langchain-core==0.1.22

# Vector Stores
pinecone-client==3.0.2
chromadb==0.4.22

# Data Processing
pandas==2.2.0
numpy==1.26.3
tiktoken==0.5.2

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
lxml==5.1.0

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.3

# Validation & Config
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Logging & Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==1.40.0

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Utilities
tenacity==8.2.3
cachetools==5.3.2
python-dateutil==2.8.2

# Security & Rate Limiting
slowapi==0.1.9
secure==0.3.0
email-validator==2.1.0

# Input Validation
bleach==6.1.0

